Updated per PR comments
